requirements.txt
pandas
openai
python-dotenv
python-dotenv
openai
rq
redis
